FLOSSmole plugin: Initial setup and commented intentions
I'm working to understand the plugin system, towards adding metrics
about software.  Clearly downloads, releases and activity percentiles
are problematic metrics, but with the right limitations (see
publications on Open Source Success measures at
http://james.howison.name) could be useful.

FLOSSmole has a straight mysql access method, so my thinking is that
the Plugin.py in this folder calls out over the network to MySQL
directly and creates the relevant json structure for the total-impact
API.  Hopefully MySQL access from your server is not problematic.

Currently all I'm doing in this commit is indicating where I think
things go in the Plugin, I'd love to feedback on that.